+++ lib/features/net_worth/view/screens/add_asset_account_screen.dart
@@ -99,7 +99,7 @@ class _AddAssetAccountScreenState extends State<AddAssetAccountScreen> {
 
                           if (value != null &&
                               _assetNameController.text.isNotEmpty) {
-                            final assetValue = value.toDouble();
+                            final assetValue = value;
                             final assetEntity = AssetEntity(
                               name: _assetNameController.text,
                               amount: assetValue,
